red circle OK!

+++ app/src/main/java/com/example/fuckingbird/element/Tori.java
@@ -11,7 +11,7 @@ import com.example.fuckingbird.graphic.Iro;
 public class Tori {
 
     private static final float X = 100;
-    private static final float HANKEI = 50;
+    private static final int HANKEI = 50;
     private static final Paint RED = Iro.getToriNoIro();
     private float takasa;
 
@@ -20,8 +20,6 @@ public class Tori {
     }
 
     public void paint(Canvas canvas) {
-
-        canvas.drawCircle(X, takasa, HANKEI, RED);
+        canvas.drawCircle(X,takasa,HANKEI,RED);
     }
-
 }
